Remove post_header connected component and move presentation component to post_view folder
