+++ components/post_view/post/post.jsx
@@ -13,7 +13,7 @@ import * as PostUtils from 'utils/post_utils.jsx';
 import * as Utils from 'utils/utils.jsx';
 
 import PostBody from 'components/post_view/post_body';
-import PostHeader from 'components/post_view/post_header/post_header.jsx';
+import PostHeader from 'components/post_view/post_header.jsx';
 import ProfilePicture from 'components/profile_picture.jsx';
 
 const ActionTypes = Constants.ActionTypes;
